backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
redis==5.0.8
celery==5.4.0
python-dotenv==1.0.1
httpx==0.27.2
praw==7.7.1
google-play-scraper==1.2.3
